View, create, edit, delete categories
- Can now view, create, edit and delete categories. Only admins have these permissions.
- Can now change the status of a ticket if the user is an admin.
- Changed the column 'title' to 'name' in the categories table.
- Added routes for editing assigned_id and status columns on tickets.
- Removed Assign.vue and instead have the different edit forms for ticket be checked by current route name.

+++ app/Http/Controllers/CategoryController.php
@@ -5,32 +5,41 @@ namespace App\Http\Controllers;
 use App\Http\Requests\CategoryRequest;
 use App\Http\Resources\CategoryResource;
 use App\Models\Category;
+use Illuminate\Support\Facades\Gate;
 
 class CategoryController extends Controller
 {
     public function index()
     {
+        Gate::authorize('viewAny');
+
         return CategoryResource::collection(Category::all());
     }
 
     public function store(CategoryRequest $request)
     {
-        $category = Category::create($request->validated());
+        Gate::authorize('create');
 
-        $categories = Category::all();
-        return CategoryResource::collection($categories);
+        $data = $request->validated();
+
+        $category = Category::create($data);
+
+        return new CategoryResource($category);
     }
 
     public function update(CategoryRequest $request, Category $category)
     {
+        Gate::authorize('update', $category);
+
         $category->update($request->validated());
 
-        $categories = Category::all();
-        return CategoryResource::collection($categories);
+        return new CategoryResource($category);
     }
 
     public function destroy(Category $category)
     {
+        Gate::authorize('delete', $category);
+
         $category->delete();
         
         return response()->json(['message' => 'Categorie succesvol verwijderd']);
